requirements.txt
Pillow>=9.0.0
qrcode>=7.3.1 